backend/requirements.txt
requests
beautifulsoup4
firebase-admin
python-dotenv
openai
fastapi
uvicorn
